Move ColorScale and Map component in lib
- Convert the following components to typescript:
    - Map/index.tsx
    - Map/OlMap/index.tsx
    - Map/HeatmapLayer/index.tsx
    - ColorScale/index.tsx
- NOTE: Both lib and storybook are broken
- Add dependencies for Map component
- Remove eslint check for require-default-props
