+++ lib/package.json
@@ -39,7 +39,16 @@
         "@babel/runtime-corejs3": "^7.22.6",
         "@knight-lab/timelinejs": "^3.9.3",
         "@togglecorp/fujs": "^2.1.1",
-        "react-icons": "^5.0.1"
+        "d3": "^7.8.5",
+        "d3-scale": "^4.0.2",
+        "d3-scale-chromatic": "^3.0.0",
+        "file-saver": "^2.0.5",
+        "html-to-image": "^1.11.11",
+        "ol": "^7.3.0",
+        "ol-ext": "^4.0.10",
+        "react-icons": "^5.0.1",
+        "statsbreaks": "^1.0.4",
+        "webfontloader": "^1.6.28"
     },
     "peerDependencies": {
         "react": "^18.2.0",
@@ -56,6 +65,7 @@
         "@rollup/plugin-eslint": "^9.0.4",
         "@rollup/plugin-image": "^3.0.3",
         "@rollup/plugin-node-resolve": "^15.1.0",
+        "@types/geojson": "^7946.0.14",
         "@types/jest": "^29.5.2",
         "@types/node": "^20.4.1",
         "@types/react": "^18.2.14",
