+++ lib/src/components/ColorScale/styles.module.css
