+++ lib/src/components/Map/assets/logos/deep_small.png
