+++ lib/src/components/Map/assets/logos/immap.png
