+++ lib/src/components/Map/assets/logos/immap_strapline.png
